+++ doomgeneric/p_floor.c
@@ -508,13 +508,13 @@ EV_BuildStairs
 		    continue;
 					
 		tsec = (sec->lines[i])->frontsector;
-		newsecnum = tsec-sectors;
+		newsecnum = (int) (tsec-sectors);
 		
 		if (secnum != newsecnum)
 		    continue;
 
 		tsec = (sec->lines[i])->backsector;
-		newsecnum = tsec - sectors;
+		newsecnum = (int) (tsec - sectors);
 
 		if (tsec->floorpic != texture)
 		    continue;
